[#14]Refactor: Carousel 수동 전환, recoil 제거, 페이지네이션 파일 위치 정리
