+++ package.json
@@ -30,7 +30,6 @@
     "react-native-splash-screen": "^3.3.0",
     "react-native-svg": "^15.14.0",
     "react-native-web": "^0.21.2",
-    "recoil": "^0.7.7",
     "zustand": "^5.0.8"
   },
   "devDependencies": {
